pyproject.toml
[build-system]
requires = ["setuptools>=77, <80.9", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "graphsense-lib"
description = "Graphsense backend lib and automation cli"
authors = [{name = "Iknaio Cryptoasset Analytics GmbH", email = "contact@iknaio.com"}]
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Topic :: Utilities",
]
requires-python = ">=3.10, <3.14"
dependencies = [
    'setuptools>=80.0.0, <80.9',
    "filelock>=3.8.0",
    "click>=8.0.3",
    "pandas>=2.3.3",
    "methodtools>=0.4",
    "simplejson>=3.17.6",
    "goodconf[yaml]>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0,<2.13.0",
    "requests>=2.32.5",
    "parsy>=2.0, <3.0",
    "rich>=12.6.0",
    "cashaddress>=1.0.6",
    "base58>=2.1",
    "bitarray>=2.9",
    "eth-event>=1.2.3",
    "eth-hash>=0.3.0",
    "cassandra-driver>=3.27.0",
    "async-lru>=2.0.4",
    "tenacity>=8.2.3",
    "coincurve>=21.0.0",
    "bech32>=1.2.0",
    "httpx>=0.28.1",
    "ecdsa>=0.13",
    "eth-account>=0.13.7",
]
dynamic = ["version"]
keywords=["graphsense"]

[project.readme]
file = "README.md"
content-type = "text/markdown; charset=UTF-8; variant=GFM"
# Add here related links, for example:

[project.urls]
Homepage = "https://graphsense.github.io/"
Source = "https://github.com/graphsense/graphsense-lib"
Changelog = "https://github.com/graphsense/graphsense-lib/blob/master/CHANGELOG.md"
Tracker = "https://github.com/graphsense/graphsense-lib/issues"
Download = "https://github.com/graphsense"
# Twitter = https://twitter.com/graphsense
# Change if running only on Windows, Mac or Linux (comma-separated)
# Add here all kinds of additional classifiers as defined under
# https://pypi.org/classifiers/


[dependency-groups]
# Core testing dependencies
testing = [
    "setuptools>=80.0.0, <80.9",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "testcontainers[cassandra]>=4.9.1",
    "testcontainers[postgres]>=4.9.1",
    "pytest-vcr",
    "vcrpy",
    "pytest-asyncio>=0.26.0",
]

# Development dependencies
dev = [
    "setuptools>=80.0.0, <80.9",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.1",
    "pre-commit>=4.1.0",
    "setuptools_scm>=8.1.0",
    "pdbpp>=0.10.3",
    "grpcio-tools>= 1.49.0",
    "testcontainers[cassandra]>=4.9.1",
    "testcontainers[postgres]>=4.9.1",
    "pytest-vcr",
    "vcrpy",
    "pytest-asyncio>=0.26.0",
    "build>=1.2.2.post1", # dependency only for restriction for vcrpy
    "httpx<1.0.0",
    "ty>=0.0.1a20",
]
[project.optional-dependencies]
# Core ingest infrastructure dependencies
ingest = [
    "graphsense-ethereum-etl>=2.4.6",
    "chainside-btcpy>=0.6",
    "graphsense-bitcoin-etl>=1.5.5",
    "lz4>=4.3.2",
    "pyarrow>=23.0.0",
    "deltalake>=1.4.2",
    "duckdb>=1.0.0",
    "grpcio>=1.49.0",
]
conversions = [
    "networkx>=3.2.1",
    "matplotlib",
]

tagstore = [
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.3",
    "colorama>=0.4.6",
    "fastapi[all]>=0.115.0",
    "httpx<1.0.0",
    "gitpython>=3.1",
    "giturlparse>=0.10",
    "lz4>=4.3.2",
    "python-dotenv>=1.0.1",
    "pyyaml-include~=1.3",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "anytree>=2.9.0",
]

tagpacks = [
    "graphsense-lib[tagstore]",
    "rapidyaml>=0.10.0",
    "rapidfuzz>=3.13.0",
]

web = [
    "graphsense-lib[conversions,tagstore]",
    "asgi-lifespan>=2.1.0",
    "fastapi>=0.115.0",
    "python-dateutil>=2.9.0",
    "redis>=5.0.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0",
]

all = [
    "graphsense-lib[ingest]",
    "graphsense-lib[conversions]",
    "graphsense-lib[tagpacks]",
    "graphsense-lib[web]",
]

[project.scripts]
graphsense-cli = "graphsenselib.cli:main.main"

[tool.setuptools]
zip-safe = true
include-package-data = true
package-dir = {"" = "src"}
# Require a min/specific Python version (comma-separated conditions)
# Add here dependencies of your project (line-separated), e.g. requests>=2.2,<3.0.
# Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
# new major versions. This works if the required packages follow Semantic Versioning.
# For more information, check out https://semver.org/.
platforms = ["any"]


[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"


[tool.setuptools.packages.find]
where = ["src"]
exclude = [
    "tests",
    "src/grapsenselib/ingest/tron/protos",
]
namespaces = true


[tool.pytest.ini_options]
# Specify command line options as you would do when invoking pytest directly.
# e.g. --cov-report html (or xml) for html/xml output or --junitxml junit.xml
# in order to write a coverage file that can be read by Jenkins.
# CAUTION: --cov flags may prohibit setting breakpoints while debugging.
# Comment those flags to avoid this pytest issue.
norecursedirs = [
    "dist",
    "build",
    "tests/regressions",
]
testpaths = ["tests"]
# Use pytest markers to select/deselect specific tests
markers = [
    "slow: mark tests as slow (deselect with '-m \"not slow\"')",
    "vcr:",
]
asyncio_mode = "auto"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "openapi_server",
    "tests/test_*.py",
    "tests/web",
    "clients",
    "setup.py",
    "src/graphsenselib/ingest/tron/grpc*",
    "src/tagpack",
    "src/tagstore",
    "tests/regressions"
]

[tool.ruff.lint]
extend-select = ["T201"]

[tool.ty.src]
include = ["src/graphsenselib/config/**",
          "src/graphsenselib/schema/**",
          "src/graphsenselib/errors/**",
          "src/graphsenselib/datatypes/**",
          "src/graphsenselib/utils/**",
          "tests/config/**",
          "tests/schema/**",
          "tests/errors/**",
          "tests/datatypes/**",
          "tests/utils/**"
]
exclude = ["src/graphsenselib/ingest/tron/grpc/**"]

[tool.ty.rules]
unused-ignore-comment = "error"
unresolved-import = "warn"
